Restore geiser server for STklos (#8629)
Geiser is a collection of Emacs modes that works with different Scheme
implementations. Each implementation (for each specific Scheme system)
needs one Emacs Lisp frontend, and one Scheme backend -- and the
backend for STklos is the file `geiser.stk`, which is needed for
geiser-stklos to work (otherwise the package just won't work).

Version 1.4 of geiser-stklos in MELPA had `geiser.stk`, but this file
was removed in commit ad320d60e2c95881f31628c19ad3b9ece7e3d165.

This commit restores that file.