Ignore texi files that are known to be broken
Building these manuals fails, which isn't fatal but just means that
we don't distribute them.  This commit doesn't change that, it just
means that we don't even try to build these manuals.

The benefit of that is that we won't see errors messages about these
known upstream issues when building packages, and that helps focusing
on other error messages, which may be serious and need to be addressed
instead of being hidden in a see of benign issues.

The cost is, that after addressing the underlying issue, upstream
maintainers also have to ask be to remove this kludge.

Tracked in #8690.