Update url of ob-powershell (#8768)
Since #7932, @rkinggen has already accepted changes from @MoisMoshev and the current url remains 3 commits behind for almost a whole year!

This is *not* a change of ownership. @rkinggen is the original author and did not step down as the maintainer. We simply used to import from a fork by mistake. -- Jonas